#21 Refactor
